Move as is file specs to codegen
