+++ generators/swift/base/src/AsIs.ts
@@ -5,17 +5,6 @@ import { RelativeFilePath } from "@fern-api/fs-utils";
 import { swift } from "@fern-api/swift-codegen";
 
 /**
- * Configuration specification for a static Swift file that gets included as-is in the generated SDK.
- * This serves as the raw configuration that gets transformed into a fully resolved {@link AsIsFileDefinition}
- * during the build process.
- */
-interface AsIsFileSpec<SymbolName extends string> {
-    relativePathToDir: string;
-    filenameWithoutExtension: string;
-    symbols?: { name: SymbolName; shape: swift.TypeSymbolShape }[];
-}
-
-/**
  * A fully resolved definition of a static Swift file, ready for use in codegen.
  */
 export interface AsIsFileDefinition {
@@ -45,138 +34,9 @@ export interface AsIsFileDefinition {
 }
 
 /**
- * Registry of all static Swift file specifications.
- *
- * This constant defines the complete catalog of pre-written Swift files that can be
- * included in generated SDKs. Each entry maps a unique identifier to a file specification
- * containing the file's intended path and exported symbols.
- */
-const SourceAsIsFileSpecs = {
-    // Core/Networking
-    Http: {
-        relativePathToDir: "Core/Networking",
-        filenameWithoutExtension: "HTTP",
-        symbols: [{ name: "HTTP", shape: { type: "enum-container" } }]
-    },
-    HttpClient: {
-        relativePathToDir: "Core/Networking",
-        filenameWithoutExtension: "HTTPClient",
-        symbols: [{ name: "HTTPClient", shape: { type: "class" } }]
-    },
-    MultipartFormData: {
-        relativePathToDir: "Core/Networking",
-        filenameWithoutExtension: "MultipartFormData",
-        symbols: [{ name: "MultipartFormData", shape: { type: "class" } }]
-    },
-    MultipartFormDataConvertible: {
-        relativePathToDir: "Core/Networking",
-        filenameWithoutExtension: "MultipartFormDataConvertible",
-        symbols: [{ name: "MultipartFormDataConvertible", shape: { type: "protocol" } }]
-    },
-    MultipartFormField: {
-        relativePathToDir: "Core/Networking",
-        filenameWithoutExtension: "MultipartFormField",
-        symbols: [{ name: "MultipartFormField", shape: { type: "enum-with-associated-values" } }]
-    },
-    QueryParameter: {
-        relativePathToDir: "Core/Networking",
-        filenameWithoutExtension: "QueryParameter",
-        symbols: [{ name: "QueryParameter", shape: { type: "enum-with-associated-values" } }]
-    },
-    // Core/Serde
-    DecoderPlusAdditionalProperties: {
-        relativePathToDir: "Core/Serde",
-        filenameWithoutExtension: "Decoder+AdditionalProperties"
-    },
-    EncodableValue: {
-        relativePathToDir: "Core/Serde",
-        filenameWithoutExtension: "EncodableValue",
-        symbols: [{ name: "EncodableValue", shape: { type: "struct" } }]
-    },
-    EncoderPlusAdditionalProperties: {
-        relativePathToDir: "Core/Serde",
-        filenameWithoutExtension: "Encoder+AdditionalProperties"
-    },
-    JSONEncoderPlusEncodableValue: {
-        relativePathToDir: "Core/Serde",
-        filenameWithoutExtension: "JSONEncoder+EncodableValue"
-    },
-    KeyedDecodingContainerPlusNullable: {
-        relativePathToDir: "Core/Serde",
-        filenameWithoutExtension: "KeyedDecodingContainer+Nullable"
-    },
-    KeyedEncodingContainerPlusNullable: {
-        relativePathToDir: "Core/Serde",
-        filenameWithoutExtension: "KeyedEncodingContainer+Nullable"
-    },
-    Serde: {
-        relativePathToDir: "Core/Serde",
-        filenameWithoutExtension: "Serde",
-        symbols: [{ name: "Serde", shape: { type: "class" } }]
-    },
-    StringKey: {
-        relativePathToDir: "Core/Serde",
-        filenameWithoutExtension: "StringKey",
-        symbols: [{ name: "StringKey", shape: { type: "struct" } }]
-    },
-
-    // Core
-    CalendarDate: {
-        relativePathToDir: "Core",
-        filenameWithoutExtension: "CalendarDate",
-        symbols: [{ name: "CalendarDate", shape: { type: "struct" } }]
-    },
-    DataPlusString: {
-        relativePathToDir: "Core",
-        filenameWithoutExtension: "Data+String"
-    },
-    StringPlusUrlEncoding: {
-        relativePathToDir: "Core",
-        filenameWithoutExtension: "String+URLEncoding"
-    },
-
-    // Public
-    APIErrorResponse: {
-        relativePathToDir: "Public",
-        filenameWithoutExtension: "APIErrorResponse",
-        symbols: [{ name: "APIErrorResponse", shape: { type: "struct" } }]
-    },
-    ClientConfig: {
-        relativePathToDir: "Public",
-        filenameWithoutExtension: "ClientConfig",
-        symbols: [{ name: "ClientConfig", shape: { type: "class" } }]
-    },
-    ClientError: {
-        relativePathToDir: "Public",
-        filenameWithoutExtension: "ClientError",
-        symbols: [{ name: "ClientError", shape: { type: "enum-with-associated-values" } }]
-    },
-    FormFile: {
-        relativePathToDir: "Public",
-        filenameWithoutExtension: "FormFile",
-        symbols: [{ name: "FormFile", shape: { type: "struct" } }]
-    },
-    JsonValue: {
-        relativePathToDir: "Public",
-        filenameWithoutExtension: "JSONValue",
-        symbols: [{ name: "JSONValue", shape: { type: "enum-with-associated-values" } }]
-    },
-    Nullable: {
-        relativePathToDir: "Public",
-        filenameWithoutExtension: "Nullable",
-        symbols: [{ name: "Nullable", shape: { type: "enum-with-associated-values" } }]
-    },
-    RequestOptions: {
-        relativePathToDir: "Public",
-        filenameWithoutExtension: "RequestOptions",
-        symbols: [{ name: "RequestOptions", shape: { type: "struct" } }]
-    }
-} satisfies Record<string, AsIsFileSpec<swift.AsIsSymbolName>>;
-
-/**
  * Union type of all available static file identifiers.
  */
-export type SourceAsIsFileId = keyof typeof SourceAsIsFileSpecs;
+export type SourceAsIsFileId = keyof typeof swift.SourceAsIsFileSpecs;
 
 /**
  * Mapped type that provides strongly-typed access to all static file definitions.
@@ -214,8 +74,9 @@ export const SourceAsIsFiles = createSourceAsIsFiles();
 function createSourceAsIsFiles(): SourceAsIsFileDefinitionsById {
     const result = {} as SourceAsIsFileDefinitionsById;
 
-    for (const [key, spec] of entries(SourceAsIsFileSpecs)) {
-        const { relativePathToDir, filenameWithoutExtension, symbols } = spec as AsIsFileSpec<swift.AsIsSymbolName>;
+    for (const [key, spec] of entries(swift.SourceAsIsFileSpecs)) {
+        const { relativePathToDir, filenameWithoutExtension, symbols } =
+            spec as swift.AsIsFileSpec<swift.AsIsSymbolName>;
         result[key] = {
             filenameWithoutExtension,
             directory: RelativeFilePath.of(relativePathToDir),
@@ -237,7 +98,7 @@ const TestAsIsFileSpecs = {
         filenameWithoutExtension: "WireStub",
         symbols: [{ name: "WireStub", shape: { type: "class" } }]
     }
-} satisfies Record<string, AsIsFileSpec<string>>;
+} satisfies Record<string, swift.AsIsFileSpec<string>>;
 
 export type TestAsIsFileId = keyof typeof TestAsIsFileSpecs;
 
@@ -251,7 +112,7 @@ function createTestAsIsFiles(): TestAsIsFileDefinitionsById {
     const result = {} as TestAsIsFileDefinitionsById;
 
     for (const [key, spec] of entries(TestAsIsFileSpecs)) {
-        const { relativePathToDir, filenameWithoutExtension, symbols } = spec as AsIsFileSpec<string>;
+        const { relativePathToDir, filenameWithoutExtension, symbols } = spec as swift.AsIsFileSpec<string>;
         result[key] = {
             filenameWithoutExtension,
             directory: RelativeFilePath.of(relativePathToDir),
