+++ generators/swift/codegen/src/symbol/index.ts
@@ -1,2 +1,3 @@
-export { type AsIsSymbolName, type FoundationTypeSymbolName, type SwiftTypeSymbolName, Symbol } from "./symbol";
+export { type AsIsFileSpec, type AsIsSymbolName, SourceAsIsFileSpecs } from "./as-is";
+export { type FoundationTypeSymbolName, type SwiftTypeSymbolName, Symbol } from "./symbol";
 export { SymbolRegistry, type TypeSymbolShape } from "./symbol-registry";
